# openapi.yaml
openapi: 3.0.0
info:
  title: 美妝比價平台 API
  description: 這是美妝比價平台的 API 文件
  version: 1.7.1
servers:
  - url: http://localhost:5000/
...
paths:
...
  /loginpage/login:
    post:
      tags:
        - LoginPage
      summary: (+token)表單驗證完登入
      description: 先做表單驗證，account password必填
      requestBody:
        description: 搜尋請求
        required: true
        content:
          application/json:
            schema: 
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: "success"
        "400":
          description: "Invalid request parameter"
        "404":
          description: "find no result"
...
  /gooddetail/product/{pId} :
    get:
      tags:
        - GoodDetail
      summary: 取得特定商品詳細資料
      description: 把pId傳進來
      parameters:
        - name: pId
          in: path
          required: true
          description: 商品編號
          schema:
              type: string
              example: "a0000001"
      responses:
        "200":
          description: "success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: object
                    properties:
                      pId:
                        type: string
                        maxLength: 8
                        example: "a0000001"
                      pName:
                        type: string
                        maxLength: 50
                        example: "霓淨思超爆水面膜5片-潤澤保濕"
                      brand:
                        type: string
                        maxLength: 30
                        example: "霓淨思"
                      category:
                        type: string
                        maxLength: 20
                        example: "臉部保養"
                      price:
                        type: number
                        example: 99.0
                      clickTimes:
                        type: integer
                        example: 3
                      review:
                        type: number
                        example: 4.5
        "400":
          description: "Invalid request parameter"
        "404":
          description: "find no result"
  /gooddetail/priceNow/{pId}:
   get:
     tags:
      - GoodDetail
     summary: 取得特定商品各通路價格
     description: 把pId傳進來
     parameters:
      - name: pId
        in: path
        required: true
        description: 商品編號
        schema:
            type: string
            example: "a0000001"
     responses:
      "200":
        description: "success"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Price_Now'
            examples:
              example1:
                summary: 商品 a0000001 各通路價格資訊
                value:
                  results:
                    - pId: "a0000001"
                      store: "poya"
                      updateTime: "2025-03-03T19:00:00Z"
                      storePrice: 350
                      storeDiscount: "買一送一，限時優惠"
                      storeLink: "https://www.poya.com.tw/product/a0000001"
                    - pId: "a0000001"
                      store: "watsons"
                      updateTime: "2025-03-03T19:00:00Z"
                      storePrice: 298
                      storeDiscount: "第二件 5 折"
                      storeLink: "https://www.watsons.com.tw/product/a0000001"
                    - pId: "a0000001"
                      store: "cosmed"
                      updateTime: "2025-03-03T19:00:00Z"
                      storePrice: 298
                      storeDiscount: "滿 500 送 50"
                      storeLink: "https://www.cosmed.com.tw/product/a0000001"
      "400":
        description: "Invalid request parameter"
      "404":
        description: "find no result"
  /gooddetail/productReview/{pId}:
    get:
      tags:
        - GoodDetail
      summary: 取得商品評價
      description: 根據 `pId` 查詢該商品的所有評價
      parameters:
      - name: pId
        in: path
        required: true
        description: 商品編號
        schema:
            type: string
            example: "a0000001"
      responses:
        "200":
          description: "成功取得商品評價"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GoodReview"
        "400":
          description: "請求參數錯誤 (例如 pId 未提供或格式錯誤)"
        "404":
          description: "找不到該商品的評價"
  /gooddetail/click/{pId}:
    post:
      tags:
        - GoodDetail
      summary: 進入商品頁面時，點擊次數 +1
      description: 用戶進入商品頁面時，點擊次數自動增加 1
      parameters:
        - name: pId
          in: path
          required: true
          description: 商品編號
          schema:
            type: string
            maxLength: 8
            example: "a0000001"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pId:
                  type: string
                  maxLength: 8
                  example: "a0000001"
      responses:
        "200":
          description: "點擊次數更新成功"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "點擊次數已更新"
                  clickTimes:
                    type: integer
                    example: 123
        "400":
          description: "請求錯誤 (無效的商品編號)"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "無效的商品編號"
  /gooddetail/track/id:
    post:
      tags:
        - GoodDetail
      summary: 檢查特定商品是否已關注 (改變按鈕狀態)
      description: |
        - 先檢查是否登入，未登入則提示會員登入
        - 已登入則檢查 `Client_favorites` 是否有 `cId` 和 `pId`
        - `object.length > 0`，按鈕顯示 "已關注"
        - `object.length = 0`，按鈕顯示 "加入關注"
        - 若已關注，點擊按鈕彈出確認通知，詢問是否取消關注
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                cId:
                  type: string
                  maxLength: 8
                  example: "C1234567"
                pId:
                  type: string
                  maxLength: 8
                  example: "P0000001"
      responses:
        "200":
          description: "回傳查詢關注清單結果"
          content:
            application/json:
              schema:
                type: object
                properties:
                  cId:
                    type: string
                    maxLength: 8
                    description: "會員編號"
                    example: "C1234567"
                  pId:
                    type: string
                    maxLength: 8
                    description: "商品編號"
                    example: "P0000001"
                  status:
                    type: integer
                    description: "關注狀態, 1已關注"
                    example: 1
        "400":
          description: "請求錯誤"
  /gooddetail/track:
    post:
      tags:
        - GoodDetail
      summary: 加入或取消關注商品
      description: |
        - 會員需登入（前端需提供有效的 `cId`）
        - 若該會員 (`cId`) 尚未關注該商品 (`pId`)，則將其加入 `Client_Favorites`
        - 若已關注，則從 `Client_Favorites` 中刪除該紀錄，表示取消關注
        - 當 `pId` 對應商品不存在時，回傳錯誤
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cId
                - pId
              properties:
                cId:
                  type: string
                  maxLength: 8
                  example: "C1234567"
                pId:
                  type: string
                  maxLength: 8
                  example: "P0000001"
      responses:
        "200":
          description: 已成功加入或取消關注
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "已加入追蹤"
                  cId:
                    type: string
                    example: "C1234567"
                  pId:
                    type: string
                    example: "P0000001"
                  status:
                    type: integer
                    enum: [0, 1]
                    description: 0 表示已取消追蹤，1 表示已加入追蹤
        "400":
          description: 請求參數錯誤
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "請求參數錯誤"
        "404":
          description: 商品不存在
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "無法追蹤，商品 ID P0000001 不存在"
        "500":
          description: 資料庫錯誤或伺服器內部錯誤
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "伺服器內部錯誤"
...
components:
  schemas:
...
    LoginRequest:
      type: object
      properties:
        account:
          type: string
          minLength: 8
          maxLength: 20
          example: "Root_123456"
        password:
          type: string
          minLength: 8
          maxLength: 20
          example: "Root_123456"
...
    Price_Now:
      type: object
      properties:
        prices:
          type: array
          minItems: 3
          maxItems: 3
          items:
            type: object
            properties:
              pId:
                type: string
                description: "商品編號"
                example: "a0000001"
              store:
                type: string
                description: "通路名稱"
                example: "poya"
              updateTime:
                type: string
                format: date-time
                description: "價格更新時間"
                example: "2025-03-03T19:00:00Z"
              storePrice:
                type: number
                description: "通路價格"
                example: 350
              storeDiscount:
                type: string
                maxLength: 200
                description: "優惠訊息 (最多 200 字元)"
                example: "買一送一，限時優惠"
              storeLink:
                type: string
                maxLength: 200
                description: "通路購買連結"
                example: "https://www.poya.com.tw/product/a0000001"
    GoodReview:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              pId:
                type: string
                maxLength: 8
                description: "商品編號"
                example: "P1234567"
              userName:
                type: string
                maxLength: 30
                description: "用戶名稱"
                example: "小明"
              date:
                type: string
                format: date-time
                description: "評價日期"
                example: "2025-03-04T12:00:00Z"
              rating:
                type: number
                format: float
                description: "評價星級 (1.0 - 5.0)"
                example: 4.5
              reviewText:
                type: string
                description: "評價內容"
                example: "這產品真的很好用！會回購～"